Improve unit test AutomaticallyFixTooLongAETitles
- Increase timeout to reduce flakiness
- Actually verify the AE title was truncated correctly by capturing the incoming association on the server
